refactor: split doctor next steps
